Commented out heavily broken Sudoku object
I'm not sure what happened, but I discovered an instance where the
Sudoku object was behaving nowhere near intended. I don't really know if
this is a problem within the Matrix object either, so more extensive
testing of both is required.
The Sudoku object that exists only has a different __str__ function, to
test to see if the __str__ code of the original was working (it is).