Added isconsistent.py
I've had this code written since I cleaned up the isvalid() docstring,
but I made the mistake of adding it all at once with get_sec and
get_slt, so it's been sitting around in an editor for a while now.
Wanted to make sure I didn't lost it, but I don't want to use it yet
until I'm further in my unittesting. This is literally just temporarily
loose code.